pyproject.toml
[project]
name = "food-calo-estimate"
version = "0.1.0"
description = ""
authors = [
    {name = "cuongwf1711",email = "ruivalien@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "pylint (>=3.3.7,<4.0.0)",
    "psycopg (>=3.2.7,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "nplusone (>=1.0.0,<2.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "djangorestframework-camel-case (>=1.4.2,<2.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "celery (>=5.5.2,<6.0.0)"
]

[tool.poetry]
packages = [{include = "food_calo_estimate", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
